refactor _setup_log_record_factory to run on generate_logging_config() which applies to all log records
